[COMMON] Mapper에 필요한 dependency 추가

See merge request s07-final/S07P31A604!5

+++ gardenary-be/src/main/java/com/gardenary/domain/flower/repository/FlowerRepository.java
@@ -3,5 +3,5 @@ package com.gardenary.domain.flower.repository;
 import com.gardenary.domain.flower.entity.Flower;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FlowerRepository extends JpaRepository<Flower, Long> {
+public interface FlowerRepository extends JpaRepository<Flower, Integer> {
 }
