+++ gardenary-be/src/main/java/com/gardenary/domain/flower/repository/MyFlowerRepository.java
@@ -3,5 +3,5 @@ package com.gardenary.domain.flower.repository;
 import com.gardenary.domain.flower.entity.MyFlower;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MyFlowerRepository extends JpaRepository<MyFlower, Long> {
+public interface MyFlowerRepository extends JpaRepository<MyFlower, Integer> {
 }
